debug save&load, connect menu buttons

+++ src/js/WorkspaceManager.js
@@ -1,5 +1,6 @@
 import { app } from './App'
 import { Workspace } from './Objects/Workspace'
+import { uniqId } from './Tools/general'
 
 
 export class WorkspaceManager {
@@ -18,9 +19,22 @@ export class WorkspaceManager {
      * @param {Workspace} workspace
      */
     setWorkspace(workspace) {
+        if(!workspace || !workspace.id) {
+            console.error("Workspace object is not valid");
+            return;
+        }
         app.workspace = workspace;
-        //TODO: update drawAPI (zoom, scale)
-        //TODO: refresh.
+        workspace.history.updateMenuState();
+        app.refreshWindow();
+    }
+
+    /**
+     * Créer un nouvel objet Workspace
+     * @param  {String} [envName='Grandeur'] Nom de l'environnement
+     * @return {Workspace}
+     */
+    getNewWorkspace(envName = 'Grandeur') {
+        return new Workspace(app.envManager.getNewEnv(envName));
     }
 
 
@@ -30,15 +44,24 @@ export class WorkspaceManager {
 
 
     /**
-     * Sauvegarder un espace de travail dans le localStorage
+     * Sauvegarder un espace de travail dans le localStorage.
      * @param  {Workspace} workspace
      * @param  {String} [name="Unnamed"] Nom de l'espace de travail
+     * @param  {Boolean} [eraseIfExisting=true] True: si un workspace avec le
+     *  même id est déjà sauvegardé, on écrase cette sauvegarde. False: on
+     *  génère un nouvel id pour le workspace reçu.
+     * @return {String} l'id du workspace enregistré.
      */
-    saveWorkspaceToLocalStorage(workspace, name = "Unnamed") {
+    saveWorkspaceToLocalStorage(workspace, name = "Unnamed", eraseIfExisting = true) {
         let json = workspace.saveToJSON(),
+            index = this.getWorkspaceIndexInLocalStorage(workspace.id);
+        if(index == -1 || !eraseIfExisting) {
             index = parseInt(this.ls.getItem("AG_WorkspacesAmount"));
+            this.ls.setItem("AG_WorkspacesAmount", index + 1);
+            if(!eraseIfExisting && index != -1)
+                workspace.id = uniqId();
+        }
 
-        this.ls.setItem("AG_WorkspacesAmount", index + 1);
         this.ls.setItem("AG_WSList_WS" + index + "_uniqid", workspace.id);
         this.ls.setItem("AG_WSList_WS" + index + "_name", name);
         this.ls.setItem("AG_WSList_WS" + index + "_data", json);
@@ -81,20 +104,29 @@ export class WorkspaceManager {
     }
 
     /**
-     * Supprimer un workspace du localStorage
-     * @param  {[type]} uniqid id unique du workspace
+     * Récupérer l'index d'un Workspace dans le localStorage. Retourne -1 si le
+     * Workspace n'a pas été trouvé.
+     * @param  {String} uniqid l'id du workspace
+     * @return {int}           son index
      */
-    deleteWorkspaceFromLocalStorage(uniqid) {
-        //Find index:
+    getWorkspaceIndexInLocalStorage(uniqid) {
         let wsAmount = parseInt(this.ls.getItem("AG_WorkspacesAmount")),
             index = -1;
         for(let i=0; i<wsAmount; i++) {
             let val = this.ls.getItem("AG_WSList_WS" + i + "_uniqid");
             if (val == uniqid) {
-                index = i;
-                break;
+                return i;
             }
         }
+        return -1;
+    }
+
+    /**
+     * Supprimer un workspace du localStorage
+     * @param  {[type]} uniqid id unique du workspace
+     */
+    deleteWorkspaceFromLocalStorage(uniqid) {
+        let index = this.getWorkspaceIndexInLocalStorage(uniqid);
         if(index==-1) {
             console.error("Workspace not found");
             return;
@@ -120,18 +152,37 @@ export class WorkspaceManager {
 
 
     /**
-     * Récupérer un espace de travail depuis un fichier (async), puis appeler
-     * le callback (avec l'objet Workspace en paramètre)
+     * Récupérer un espace de travail depuis un fichier (async), et définir le
+     * Workspace actuel (appelle this.setWorkspace)
      * @param  {File} file Le fichier
      */
-    getWorkspaceFromFile(file, callback) {
+    setWorkspaceFromFile(file, callback) {
         const reader = new FileReader();
         reader.readAsText(file);
         reader.onload = () => {
 			let ws = new Workspace();
             ws.initFromJSON(reader.result);
-            callback(ws);
+            this.setWorkspace(ws);
+            if(callback) callback();
 		}
+    }
+
+    /**
+     * Sauvegarder un espace de travail comme fichier.
+     * @param  {Workspace} workspace              L'espace de travail
+     * @param  {String} [fileName='save.json'] Le nom du fichier
+     */
+    saveWorkspaceToFile(workspace, fileName) {
+        if(!fileName) //TODO (temporaire)
+            fileName = window.prompt("(popup temporaire) Nom du fichier: ") + '.json';
+
+        let json = workspace.saveToJSON();
+        const file = new Blob([json], { type: 'application/json' });
+        const downloader = document.getElementsByTagName("ag-tablette-app")[0]
+                            .shadowRoot.querySelector('#dataDownloader');
+        downloader.href = window.URL.createObjectURL(file);
+        downloader.download = fileName;
+        downloader.click();
     }
